+++ lib/intl_message.dart
@@ -5,6 +5,7 @@ import 'package:intl/src/plural_rules.dart' as plural_rules;
 import 'package:intl_message/src/intl/ordinal_rules.dart' as ordinal_rules;
 import 'dart:async';
 import 'src/parser.dart';
+import 'package:expressions/expressions.dart';
 
 part 'src/message.dart';